Parallelize tasks; enable early signaling connection establishment.
Opening the signaling websocket is one of the biggest sources of delay
when joining a meeting: even a fast connection to a nearby region can
take over 300ms.

This commit introduces an argument to `start` to allow a 'pre-start'
behavior to occur: call `start(true)` to begin signaling, and then call
`start()` to complete peer connection setup and media negotiation.

This change also parallelizes some start tasks by using a promise-based
mechanism to express a dependency chain.